pr5 refactoring
